Remove openDAQUtils.cmake workaround.
Older SDK versions did not correctly load this file in the installed
openDAQConfig.cmake file. This workaround was necessary when using a
previously installed (not fetched) SDK. This bug has been fixed for a
while in the SDK and the workaround isn't needed (the workaround also
causes problems in at least one corner case, with OVERRIDE_FIND_PACKAGE
and FetchContent()).